+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.9.0]
+- Switched to main branch
+- ILIAS 7 support
+- Remove generate deprecated PHP72Backport script
+- Remove generate deprecated GeneratePluginReadme
+- Remove generate ILIAS 5.4/PHP 7.0
+
 ## [2.8.6]
 - Change utils url